juis_check: send a faked serial number to vendor
As long as the '-r' option is not specified, the value read
from the device is only used as OUI (first 3 bytes) and the
last 3 bytes are built from a random source (epoch based).

At the same time it will be ensured, that the random value
is never the real data.