YourFritz: rebase commits from last two years to remove ...
... to remove any unnecessary reference to binary files in submodules.

The whole repo was rebased and some commits were squashed ... now there
are 438 commits, opposed to 511 earlier. Orphaned references were
removed and the whole repository is now much slimmer.

A clone of the whole repository without the submodules in "bin" and
"first_aid" needs  now below 7 MB - if you clone the submodules too,
additional 100 MB of data are transfered.