EVA-Discover: allow binding to local if address.
Depending on the Windows network setup, it might be necessary to bind
the UDP broadcast sender socket to an interface. By default, the limited
broadcast (255.255.255.255) will be sent on the interface with the
lowest metric route for this address.

Binding to the interface is done by (optionally) binding to the local IP
address given as `-if_address'.