bootmanager: add script to implant dynamic changes of OEM ...
... in /etc/init.d/rc.conf

A marker line with "# YF_CHANGE_OEM" is added, too. It's used to
decide, whether a firmware supports dynamic OEM values.

The anchor line to add the code was selected by intention - if a
fixed branding is set already, it's senseless to add this code,
because it doesn't cause any changes to the result.