yf_custom_environment.sh: reduce dependencies once again
... now the script uses /dev to create the character device file
for the used TFFS node and doesn't use/need a 'tmpfs' anymore.

(But a 'pipefs' is expected to exist, otherwise the execution of
piped command chains will fail.)

Meanwhile the implementation has changed so far (especially the
read loop over the lines with wanted values), that a 'tmpfs' isn't
needed any longer to store temporary files.