juis_check: improve XML_LINTER variable usage ...
it may now include options for the call and an existing '%s' within
is replaced by the file name. If no '%s' was specified, it will be
appended to the value. The first item (according to IFS rules) of
the value has to be the name of an executable and is checked using
a 'command -v <executable>' call. Only if this call succeeds, the
linter part is executed, otherwise a simple 'cat' will be used.

To show an example ... using 'sed' to insert line breaks after each
closing XML tag, 'XML_LINTER' could be used as follows:

XML_LINTER="sed -e 's|>\\(<[^/]\\)|>\\\\n\\\\t\\\\1|g'"

Because a combination of 'eval' and 'printf' statements is used to
prepare the command line of the linter, any backslash characters
have to be escaped as shown above for '\n', '\t' and '\1' - using
'\\(' is possible, but not critical, because '\(' (and '\)', too)
is not a valid escape sequence for 'printf'.