juis_check: improve features, fix some problems (#38)
* juis_check: be more precise, if values have to be read from device

* juis_check: add 'jason_boxinfo.xml' to considered info sources ...

if the script is used from FRITZ!OS

The error message, if neither of both files was found, is left untouched
by intention - it mentions always 'juis_boxinfo.xml' if neither file exists.

These changes are untested (but are looking good to me) - don't hesitate to
test them yourself on a proper device (I have no device close at hand yet).

* juis_check: improve conversion of decimal values ...

starting with a zero digit

* juis_check: now decimal conversion should work better ...

the regex was checked with:

i=0
while [ $i -lt 1000 ]
do
  printf "%04u=%s\n" "$i" $(expr "$(printf "%04u\n" "$i")" : "0*\([1-9]*[0-9]\+\)")
 i=$(( i + 1 ))
done 2>&1 \
| less

* juis_check: debug artifacts removed

* juis_check: implement new parameter 'Buildtype'

Usage help still needs to be updated.

* juis_check: 'Nonce' is not really a JUIS variable, ...

and may not be read from device. But with the earlier check of 'culture
clashes' for exclusive parameters, it has to be ignored in the first
attempt to assign/collect data for request properties.

* juis_check: improve bash detection and network error handling

* juis_check: add option to show (prettified) response on STDERR

* juis_check: update german usage screen

* juis_check: update english usage screen

* juis_check: add display of known Buildtype names ...

and shorten german 'Public' description (the removed items should be known
meanwhile).

* juis_check: update README.md ...

... it has still to be reviewed with a markdown viewer.

* juis_check: update german README.md ...

... has to be validated with a Markdown viewer, too.

* juis_check: remove rant about unwanted firmware updates ...

... it may not be solved finally, but it seems to get better meanwhile.

* juis_check: Markdown fixes for README files

* juis_check: improve XML_LINTER variable usage ...

it may now include options for the call and an existing '%s' within
is replaced by the file name. If no '%s' was specified, it will be
appended to the value. The first item (according to IFS rules) of
the value has to be the name of an executable and is checked using
a 'command -v <executable>' call. Only if this call succeeds, the
linter part is executed, otherwise a simple 'cat' will be used.

To show an example ... using 'sed' to insert line breaks after each
closing XML tag, 'XML_LINTER' could be used as follows:

XML_LINTER="sed -e 's|>\\(<[^/]\\)|>\\\\n\\\\t\\\\1|g'"

Because a combination of 'eval' and 'printf' statements is used to
prepare the command line of the linter, any backslash characters
have to be escaped as shown above for '\n', '\t' and '\1' - using
'\\(' is possible, but not critical, because '\(' (and '\)', too)
is not a valid escape sequence for 'printf'.

* juis_check: properly quote automatically appended file name