signimage: add a script to show problematic sizes of data to get ...
... signed

This script creates images of various sizes and checks, whether the
signature may be validated by check_signed_image AND by AVM's utility
tr069fwupdate.

It shows some size-related values in cases of differences ... to allow
creation of problematic images, the YF_SIGNIMAGE_SKIP_WORKAROUNDS
option is used in the first pass and in cases of an error, the same
size is used again with activated workarounds. In the second attempt
shouldn't occur errors anymore.