dtc: add new submodule with patches to 'dtc' to get ...
a standalone utility 'fitdump', based on code of 'fdtdump'.

Output of 'fitdump' contains separate files with content of kernel
and filesystem image(s), which may be processed further.

For some FRITZ!Box models it could be necessary to use the utility
recursive to dissect further embedded FIT images (marked by
'type = "avm,fit"' in the structure).

BLOBs of type 'ramdisk' are CPIO archives compressed with 'gzip',
even if 'compression = "none"' is used in their properties. For all
other types the properties seem to be valid as specified.