fitdump.sh: add a shell implementation for the same functionality
as in 'dtc/fitdump' binary

This script may be used on any architecture to dissect AVM's FIT image
file(s) into an 'its' file and the contained image BLOBs in separated
files.

Output was designated to be the same as with the binary version,
built from a pre-existing utility in the `dtc` project.

The structure of the first 72 bytes of AVM's image format is still
under investigation. As far as I know, there wasn't seen any FIT image
for a model with native BE byte order - therefore it's unknown, whether
the signature at offset 0x0 and the size value at offset 0x4 depend on
byte-order of the platform or are always stored in LE order.