fitdump.sh: create a symlink to 'filesystem.image' while
dissecting FIT image file

A new approach is used here to detect the correct entry with
SquashFS image for AVM's frontend system.

From seven devices with FIT images by AVM so far, six models
have an 'avm,kernel-args' property containing 'mtdparts_ext='
in a node with 'type = "filesystem";' and it's always the node
for the SquashFS image with the root-filesystem for the frontend
OS.

Thinking further, this makes sense ... this 'main system' is
the OS, which manages the flash memory content (incl. installing
new firmware images) and the other systems, that exist for some
models, are specialized images to support a specific piece of
hardware. Only the first part needs the additional kernel command
line parameter (which was implemented by AVM, by the way - it
doesn't exist in other systems).

This should be a better approach, instead of a 'use always the
biggest entry with 'type = "filesystem";' - even if the latter
one should be still valid, it's harder to implement, too.

The 7th model (5530) doesn't use a SquashFS image at all - it has
a RAM-disk image (a CPIO archive compressed with gzip) instead.