fitdump.sh: add a symlink to 'kernel.image', too, which ...
... entry is the right one, will be guessed by the existence of a
node named 'avm,kallsyms' for a kernel image.

It seems to be probable, that only the main system needs/uses its
own kernel symbols table - usually the other parts of a system with
more than a single OS instance will be copied to more than a single
model (e.g. AVM's OS for the 'Scorpion' WiSoC) and aren't intended
to be debugged on the 'final model', where they'll be used (at least
it's highly unpresumable to be so).