Revert "fitdump.sh: add a symlink to 'kernel.image', too, which ..."
This reverts commit b7228f263aa64e3c4c05107d427502b88747b577.

The approach tried doesn't work with 5530, 7510 and 1200AX - either
there are no 'kallsyms' nodes at all (it could occur in the future
for further models, if AVM decides, those symbols aren't needed any
longer) or more than one entry with such a node (5530).

So this path will be abandoned now ... but changes are kept in the
reverted commit for a later chance to check the idea again.