fitdump.sh: add handling of 'ramdisk' images (biggest image wins) and ...
an image-to-node-name dictionary (as a temporary hidden file)

Future checks of 'configurations' nodes will make it possible to find the file index
number (and using this value, the file name of the kernel image) - based on kernel
entry name in a configuration node.

Furthermore a hidden file (.order) will be added to each sub-directory in the (optional)
directory structure, containing the original order of sub-nodes and properties in this
node - it can be used to create a new 'image.its' file from directory structure. It's
easier to compare '.its' files, if they use the same order for their properties.