fitdump.sh: correct a flaw, where a hexadecimal four-bytes value ...
containing three bytes of printable data and a final NUL value was
taken as a three bytes string.

The result was VALID, but it looked a bit odd. Now it will be shown as a
hexadecimal 32-bit value - but some intended string data with a size of
3 characters may use this output format unintended, too.

But however ... the resulting 'image.its' file will contain the data to
reproduce the image data, that has been dumped - whatever data format will
be written to the 'its' file.